Add tests.
