+++ pdr_backend/ppss/trader_ss.py
@@ -1,8 +1,9 @@
 from typing import Dict, List, Union
 
+import ccxt
 from enforce_typing import enforce_types
 
-from pdr_backend.ppss.base_ss import SingleFeedMixin, CCXTExchangeMixin
+from pdr_backend.ppss.base_ss import CCXTExchangeMixin, SingleFeedMixin
 from pdr_backend.util.strutil import StrMixin
 
 
@@ -87,6 +88,17 @@ class TraderSS(SingleFeedMixin, StrMixin, CCXTExchangeMixin):
             amt = float(amt_s)
             d[coin] = amt
         return d
+
+    @enforce_types
+    def ccxt_exchange(self) -> ccxt.Exchange:
+        assert hasattr(self, "exchange_params")
+
+        mock = not hasattr(self, "tradetype") or self.tradetype != "livemock"
+
+        return self.feed.ccxt_exchange(
+            self.exchange_params,
+            mock=mock,
+        )
 
 
 # =========================================================================
